webscrapper/requirements.txt
selenium
beautifulsoup4
googlesearch-python
pymongo
webdriver-manager
requests
